pyproject.toml
[project]
name = "lex-helper"
version = "0.0.9"
description = "A helper library for building Amazon Lex chatbots"
authors = [
    {name = "Amazon Web Services"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>2.0.0,<3.0.0",
    "PyYAML>=6.0,<7.0",
    "boto3>=1.26.0",
    "botocore>=1.29.0"
]
license = {text = "Apache-2.0"}

[project.urls]
Homepage = "https://github.com/aws/lex-helper"
Documentation = "https://github.com/aws/lex-helper#readme"
Issues = "https://github.com/aws/lex-helper/issues"
Repository = "https://github.com/aws/lex-helper"
Changelog = "https://github.com/aws/lex-helper/blob/main/CHANGELOG.md"



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 125


[tool.ruff.lint]
# Select rule categories - mapping from flake8 configuration
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "N",    # pep8-naming
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "C90"  # mccabe complexity
]


# Ignore rules - mapping from flake8 ignore list
ignore = [
    "E203",  # whitespace before ':'
    "E266",  # too many leading '#' for block comment
    "E501",  # line too long (handled by formatter)
    "E701",  # multiple statements on one line (colon)
    "B905",  # zip() without an explicit strict= parameter
    "E722",  # do not use bare 'except'
    "E402",  # module level import not at top of file
    "E731",  # do not assign a lambda expression, use a def
    "F811",  # redefinition of unused name
    "N815",  # mixed-case variable in class scope (intentional for Lex API compatibility)
    "W293",  # blank line contains whitespace (in examples)
    "C901",  # complex structure (acceptable in examples)
]

[tool.ruff.lint.mccabe]
# Set max complexity to match flake8 configuration
max-complexity = 18

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Use spaces around operators
indent-style = "space"
# Respect magic trailing commas
skip-magic-trailing-comma = false
# Automatically detect line ending
line-ending = "auto"




[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
norecursedirs = [
    "examples/sample_airline_bot/cdk.out",
    "examples/sample_airline_bot/node_modules",
    "examples/sample_airline_bot/lambdas/fulfillment_function/tests",
    "node_modules",
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info"
]
addopts = [
    "--tb=short",
    "--strict-markers",
    "--disable-warnings"
]

[tool.pyright]
include = [
  "./lex_helper",
]


defineConstant = { DEBUG = true }


typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownParameterType = true
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnusedImport = false
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedVariable = true
reportConstantRedefinition = true
reportDeprecated = true
reportUnnecessaryContains = true
reportUnnecessaryComparison = true
reportMatchNotExhaustive = true
reportCallInDefaultInitializer = true

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.12.11",
    "pytest-cov>=6.2.1",
    "pre-commit>=3.0.0",
    "fuzzywuzzy>=0.18.0",
    "aws-lambda-powertools>=3.19.0",
    "python-levenshtein>=0.27.1",
]
